+++ src/components/FileBrowser/FileBrowser.jsx
@@ -1,19 +1,9 @@
 import { useEffect, useState, useRef, useLayoutEffect, useCallback } from "react";
 import "./FileBrowser.scss";
 
+import { TreeNode } from "./TreeNode/TreeNode";
+
 import { setDefaultCollapsed, collapseTree, selectNode, flattenTree } from "./helper";
-
-import { FileCode, ChevronRight, ChevronDown, Braces, FiletypeScss, FiletypeJs, FiletypePy} from "react-bootstrap-icons";
-import PropTypes from 'prop-types';
-import {
-    DndContext,
-    DragOverlay,
-    useDraggable,
-    useDroppable,
-} from "@dnd-kit/core";
-
-const INDENT_WIDTH = 20;
-const SELECTED_FILE_COLOR = "#00426b";
 
 /**
  * Used for creating a stable callback function.
@@ -28,69 +18,6 @@ function useEvent(fn) {
     return useCallback((...args) => {
         return fnRef.current(...args);
     }, []);
-}
-
-/**
- * Renders a single node in the file tree.
- */
-const TreeNode = ({id, node, onRowClick}) => {
-    const { attributes, listeners, setNodeRef, transform } = useDraggable({id});
-    /**
-     * Gets the appropriate icon for the node based on its type and collapsed state.
-     * @returns <JSX>
-     */
-    const getCollapsedIcon = () => {
-        if (node.collapsed) {
-            return <ChevronRight />;
-        } else {
-            return <ChevronDown />;
-        }
-    }
-
-    /**
-     * Sets the background color of the row if the node is selected.
-     */
-    const getRowStyle = () => {
-        const style = {};
-        if (node.selected) {
-            style["backgroundColor"] = SELECTED_FILE_COLOR;
-        }
-        return style;
-    }
-
-
-    /**
-     * Get the file icon based on the extension.
-     * 
-     * TODO: This can be improved by having a mapping of extensions to icons and color.
-     */
-    const getFileIcon = () => {
-        const extension = node.name.split('.').pop();
-        switch (extension) {
-            case "js":
-                return <FiletypeJs color="#f1e05a" />;
-            case "json":
-                return <Braces color="#fffb00" />;
-            case "scss":
-                return <FiletypeScss color="#f12727" />;
-            case "py":
-                return <FiletypePy color="#686affbd" />;
-            default:
-                return <FileCode color="#ffffff" />;
-        }
-    }
-
-    return (
-        <div className="file-node-row" ref={setNodeRef} {...listeners} {...attributes} style={getRowStyle()} onClick={() => onRowClick(node)}>
-            <div className="indent" style={{ width: node.level * INDENT_WIDTH + "px"}} />
-            {
-                node.type === "folder" ? 
-                <span className="center folder-icon">{getCollapsedIcon()}</span> :
-                <span className="center file-icon">{getFileIcon()}</span>
-            }
-            <span className="center file-name">{node.name}</span>
-        </div>
-    )
 }
 
 /**
